pyproject.toml
[project]
name = "py-songs-to-karaoke"
version = "0.1.0"
description = "A tool to create karaoke versions of songs with time-synced lyrics"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Chuan Jhe Huang", email = "cjhwong@gmail.com"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Sound/Audio",
]
dependencies = [
    "pygame>=2.0.0",
    "ffmpeg-python>=0.2.0",
]

[dependency-groups]
dev = [
    "ruff>=0.1.0",
    "pytest>=7.0.0",
    "pytest-cov>=5.0.0",
]

[tool.setuptools]
package-dir = {"" = "."}
packages = ["src"]

[tool.ruff]
line-length = 120
target-version = "py38"

# Allow imports relative to the "src" and root directories
src = ["src", "."]

[tool.ruff.lint]
# Enable Pyflakes ('F'), pycodestyle ('E'), isort ('I'), and other rule sets
select = ["E", "F", "I", "W", "N", "UP", "B", "C4"]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Indent with 4 spaces
indent-style = "space"
line-ending = "auto"
skip-magic-trailing-comma = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
pythonpath = [
  "src"
]
